Feature/consulta de ficha individual de animal. 
Se tomó nota del comentario sobre el archivo vacío. No fue modificado en este PR, por lo que el problema debe ser externo al mismo (posiblemente un error en la subida o un estado previo del archivo). Se revisará por separado.

+++ NutriTrack.API/DTOs/RegistroPesoDTO.cs
@@ -3,7 +3,7 @@
     public class RegistroPesoDTO
     {
         public DateTime FechaPesaje { get; set; }
-        public decimal PesoKg { get; set; }
+        public float PesoKg { get; set; }
         public string? Observaciones { get; set; }
         public int IdUsuario { get; set; }
         public int IdAnimal { get; set; }
