+++ NutriTrack.Core/Entities/RegistroPeso.cs
@@ -8,7 +8,8 @@ namespace NutriTrack.Core.Entities
     {
         public int Id { get; set; }
         public DateTime FechaPesaje { get; set; }
-        public decimal PesoKg { get; set; }
+        //se hizo un cambio de tipo de dato paso de decimal a float. Ya que double precision equivale a float en c#.
+        public float PesoKg { get; set; }
         public string? Observaciones { get; set; }
         public int IdUsuario { get; set; }
         public int IdAnimal { get; set; }
